ai_workbench_setup/ai_workbench/requirements.txt
fastapi==0.115.0
streamlit==1.39.0
transformers==4.44.2
torch==2.4.1
huggingface_hub[hf_xet]
openai==1.50.2
uvicorn==0.30.6
pyyaml==6.0.2
python-dotenv==1.0.1
rouge-score==0.1.2
plotly==5.24.1
pandas==2.2.3
pytest==8.3.3
langchain==0.3.1
chromadb==0.5.5
sentence-transformers==3.1.1
nltk==3.9.1
evaluate==0.4.3
peft==0.12.0
shap==0.46.0
PyPDF2==3.0.1
pytesseract==0.3.13
speechrecognition==3.10.4
gTTS==2.5.3
fastapi-websocket-pubsub==1.0.1
vaderSentiment==3.3.2
detoxify==0.2.2
langchain-community==0.3.1
bert_score
httpx==0.27.0
langchain-huggingface==0.1.0
langchain-chroma==0.1.4